Testing emoji database
I do not think I will use this approach going forward since it would mean a lot more data moving over servers in the long run.

We will continue to include the Emoji file in the app for now. Possibly split it into two files. One of which is just the skin tones